--- components/home/sections/manifest.js
@@ -1,3 +1,0 @@
-const manifest = () => {};
-
-export default manifest;
