+++ styles/style.css
@@ -472,7 +472,7 @@ hr.black{
 .text{        
     font-size: 16px;
     font-weight: 300;
-    color: #5f5f5f;
+    color: #252323;
     line-height: 1.8;
 }
 .text h1,
